requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.11.0
python-engineio==4.9.0
PyMuPDF==1.23.26
Pillow==10.1.0
python-dotenv==1.0.0
